More precise language, children vs siblings
"siblings under" means they could in theory be grandchildren. It's also confusing, because the fact that they are siblings has nothing to do with retrieving the Button. Only the fact that the Button is a child is of the Panel matters when discussing how to retrieve the button from the Panel.